Split the GW inference module by responsibility
dingo/gw/inference/factors.py (1673 lines) becomes three modules with
one concern each, moved verbatim:

  context.py  GWSamplerContext and the frequency-range update helper
  steps.py    the GNPE factors, SyntheticPhaseFactor, and the
              coordinate reparametrizations
  sampler.py  GWComposedSampler and its chain builders

All imports are retargeted repo-wide (pipe, result, importance
sampling, scan, tests, docs); no compatibility shim is kept, since the
module was introduced on this branch and has no external consumers.
Pure relocation -- no code changes.
